Make all new types private
