+++ ledger/ledger-api-client/src/main/scala/com/digitalasset/ledger/client/services/acs/ActiveContractSetClient.scala
@@ -3,50 +3,22 @@
 
 package com.daml.ledger.client.services.acs
 
-import akka.NotUsed
-import akka.stream.scaladsl.{Keep, Source}
+import akka.stream.scaladsl.Source
 import com.daml.grpc.adapter.ExecutionSequencerFactory
-import com.daml.grpc.adapter.client.akka.ClientAdapter
 import com.daml.ledger.api.domain.LedgerId
 import com.daml.ledger.api.v1.active_contracts_service.ActiveContractsServiceGrpc.ActiveContractsServiceStub
-import com.daml.ledger.api.v1.active_contracts_service.{
-  GetActiveContractsRequest,
-  GetActiveContractsResponse,
-}
+import com.daml.ledger.api.v1.active_contracts_service.GetActiveContractsResponse
 import com.daml.ledger.api.v1.transaction_filter.TransactionFilter
-import com.daml.ledger.client.LedgerClient
-import com.daml.util.akkastreams.ExtractMaterializedValue
-import scalaz.syntax.tag._
 
 import scala.concurrent.Future
 
-object ActiveContractSetClientWithoutLedgerId {
-
-  private val extractOffset =
-    new ExtractMaterializedValue[GetActiveContractsResponse, String](r =>
-      if (r.offset.nonEmpty) Some(r.offset) else None
-    )
-
-}
-
-sealed class ActiveContractSetClientWithoutLedgerId(service: ActiveContractsServiceStub)(implicit
+final class ActiveContractSetClient(
+    ledgerId: LedgerId,
+    service: ActiveContractsServiceStub,
+)(implicit
     esf: ExecutionSequencerFactory
 ) {
-
-  import ActiveContractSetClientWithoutLedgerId.extractOffset
-
-  private def request(filter: TransactionFilter, verbose: Boolean, ledgerIdToUse: LedgerId) =
-    GetActiveContractsRequest(
-      ledgerIdToUse.unwrap,
-      Some(filter),
-      verbose,
-    )
-
-  private def activeContractSource(
-      request: GetActiveContractsRequest,
-      token: Option[String],
-  ): Source[GetActiveContractsResponse, NotUsed] =
-    ClientAdapter.serverStreaming(request, LedgerClient.stub(service, token).getActiveContracts)
+  private val it = new withoutledgerid.ActiveContractSetClient(service)
 
   /** Returns a stream of GetActiveContractsResponse messages. The materialized value will
     * be resolved to the offset that can be used as a starting offset for streaming transactions
@@ -58,31 +30,6 @@ sealed class ActiveContractSetClientWithoutLedgerId(service: ActiveContractsServ
       filter: TransactionFilter,
       verbose: Boolean = false,
       token: Option[String] = None,
-      ledgerIdToUse: LedgerId,
   ): Source[GetActiveContractsResponse, Future[String]] =
-    activeContractSource(request(filter, verbose, ledgerIdToUse), token).viaMat(extractOffset)(
-      Keep.right
-    )
-}
-
-final class ActiveContractSetClient(
-    ledgerId: LedgerId,
-    service: ActiveContractsServiceStub,
-)(implicit
-    esf: ExecutionSequencerFactory
-) extends ActiveContractSetClientWithoutLedgerId(service) {
-
-  /** Returns a stream of GetActiveContractsResponse messages. The materialized value will
-    * be resolved to the offset that can be used as a starting offset for streaming transactions
-    * via the transaction service.
-    * If the stream completes before the offset can be set, the materialized future will
-    * be failed with an exception.
-    */
-  override def getActiveContracts(
-      filter: TransactionFilter,
-      verbose: Boolean = false,
-      token: Option[String] = None,
-      ledgerIdToUse: LedgerId = ledgerId,
-  ): Source[GetActiveContractsResponse, Future[String]] =
-    super.getActiveContracts(filter, verbose = verbose, token = token, ledgerId)
+    it.getActiveContracts(filter, verbose = verbose, token = token, ledgerId)
 }
