+++ ledger/ledger-api-client/src/main/scala/com/digitalasset/ledger/client/services/commands/CommandClient.scala
@@ -5,22 +5,14 @@ package com.daml.ledger.client.services.commands
 
 import akka.NotUsed
 import akka.stream.Materializer
-import akka.stream.scaladsl.{Flow, Keep, Sink, Source}
-import com.codahale.metrics.Counter
+import akka.stream.scaladsl.{Flow, Source}
 import com.daml.grpc.adapter.ExecutionSequencerFactory
 import com.daml.ledger.api.domain.LedgerId
 import com.daml.ledger.api.v1.command_completion_service.CommandCompletionServiceGrpc.CommandCompletionServiceStub
-import com.daml.ledger.api.v1.command_completion_service.{
-  CompletionEndRequest,
-  CompletionEndResponse,
-  CompletionStreamRequest,
-}
+import com.daml.ledger.api.v1.command_completion_service.CompletionEndResponse
 import com.daml.ledger.api.v1.command_submission_service.CommandSubmissionServiceGrpc.CommandSubmissionServiceStub
 import com.daml.ledger.api.v1.command_submission_service.SubmitRequest
-import com.daml.ledger.api.v1.commands.Commands.DeduplicationPeriod
 import com.daml.ledger.api.v1.ledger_offset.LedgerOffset
-import com.daml.ledger.api.validation.CommandsValidator
-import com.daml.ledger.client.LedgerClient
 import com.daml.ledger.client.configuration.CommandClientConfiguration
 import com.daml.ledger.client.services.commands.CommandTrackerFlow.Materialized
 import com.daml.ledger.client.services.commands.tracker.CompletionResponse.{
@@ -28,29 +20,36 @@ import com.daml.ledger.client.services.commands.tracker.CompletionResponse.{
   CompletionSuccess,
 }
 import com.daml.util.Ctx
-import com.daml.util.akkastreams.MaxInFlight
-import com.google.protobuf.duration.Duration
 import com.google.protobuf.empty.Empty
 import org.slf4j.{Logger, LoggerFactory}
 import scalaz.syntax.tag._
 
-import scala.concurrent.{ExecutionContext, ExecutionContextExecutor, Future}
+import scala.concurrent.{ExecutionContext, Future}
 import scala.util.Try
 
 /** Enables easy access to command services and high level operations on top of them.
   *
   * @param commandSubmissionService gRPC service reference.
   * @param commandCompletionService gRPC service reference.
+  * @param ledgerId                 Will be applied to submitted commands.
   * @param applicationId            Will be applied to submitted commands.
   * @param config                   Options for changing behavior.
   */
-sealed class CommandClientWithoutLedgerId(
+final class CommandClient(
     commandSubmissionService: CommandSubmissionServiceStub,
     commandCompletionService: CommandCompletionServiceStub,
+    val ledgerId: LedgerId,
     applicationId: String,
     config: CommandClientConfiguration,
     logger: Logger = LoggerFactory.getLogger(getClass),
 )(implicit esf: ExecutionSequencerFactory) {
+  private val it = new withoutledgerid.CommandClient(
+    commandSubmissionService,
+    commandCompletionService,
+    applicationId,
+    config,
+    logger,
+  )
 
   type TrackCommandFlow[Context] =
     Flow[
@@ -68,18 +67,7 @@ sealed class CommandClientWithoutLedgerId(
   def submitSingleCommand(
       submitRequest: SubmitRequest,
       token: Option[String] = None,
-  ): Future[Empty] =
-    submit(token)(submitRequest)
-
-  private def submit(token: Option[String])(submitRequest: SubmitRequest): Future[Empty] = {
-    logger.debug(
-      "Invoking grpc-submission on commandId={}",
-      submitRequest.commands.map(_.commandId).getOrElse("no-command-id"),
-    )
-    LedgerClient
-      .stub(commandSubmissionService, token)
-      .submit(submitRequest)
-  }
+  ): Future[Empty] = it.submitSingleCommand(submitRequest, token)
 
   /** Submits and tracks a single command. High frequency usage is discouraged as it causes a dedicated completion
     * stream to be established and torn down.
@@ -87,23 +75,10 @@ sealed class CommandClientWithoutLedgerId(
   def trackSingleCommand(
       submitRequest: SubmitRequest,
       token: Option[String] = None,
-      ledgerIdToUse: LedgerId,
   )(implicit
       mat: Materializer
-  ): Future[Either[CompletionFailure, CompletionSuccess]] = {
-    implicit val executionContext: ExecutionContextExecutor = mat.executionContext
-    val commands = submitRequest.getCommands
-    val effectiveActAs = CommandsValidator.effectiveSubmitters(commands).actAs
-    for {
-      tracker <- trackCommandsUnbounded[Unit](effectiveActAs.toList, token, ledgerIdToUse)
-      result <- Source
-        .single(Ctx.unit(CommandSubmission(commands)))
-        .via(tracker)
-        .runWith(Sink.head)
-    } yield {
-      result.value
-    }
-  }
+  ): Future[Either[CompletionFailure, CompletionSuccess]] =
+    it.trackSingleCommand(submitRequest, token, ledgerId)
 
   /** Tracks the results (including timeouts) of incoming commands.
     * Applies a maximum bound for in-flight commands which have been submitted, but not confirmed through command completions.
@@ -115,18 +90,9 @@ sealed class CommandClientWithoutLedgerId(
   def trackCommands[Context](
       parties: Seq[String],
       token: Option[String] = None,
-      ledgerIdToUse: LedgerId,
   )(implicit
       ec: ExecutionContext
-  ): Future[TrackCommandFlow[Context]] = {
-    for {
-      tracker <- trackCommandsUnbounded[Context](parties, token, ledgerIdToUse)
-    } yield {
-      // The counters are ignored on the client
-      MaxInFlight(config.maxCommandsInFlight, new Counter, new Counter)
-        .joinMat(tracker)(Keep.right)
-    }
-  }
+  ): Future[TrackCommandFlow[Context]] = it.trackCommands(parties, token, ledgerId)
 
   /** Tracks the results (including timeouts) of incoming commands.
     *
@@ -137,188 +103,23 @@ sealed class CommandClientWithoutLedgerId(
   def trackCommandsUnbounded[Context](
       parties: Seq[String],
       token: Option[String] = None,
-      ledgerIdToUse: LedgerId,
   )(implicit
       ec: ExecutionContext
-  ): Future[TrackCommandFlow[Context]] =
-    for {
-      ledgerEnd <- getCompletionEnd(token, ledgerIdToUse)
-    } yield {
-      partyFilter(parties.toSet)
-        .via(commandUpdaterFlow[Context](ledgerIdToUse))
-        .viaMat(
-          CommandTrackerFlow[Context, NotUsed](
-            commandSubmissionFlow = CommandSubmissionFlow[(Context, String)](
-              submit(token),
-              config.maxParallelSubmissions,
-            ),
-            createCommandCompletionSource =
-              offset => completionSource(parties, offset, token, ledgerIdToUse),
-            startingOffset = ledgerEnd.getOffset,
-            maximumCommandTimeout = config.defaultDeduplicationTime,
-          )
-        )(Keep.right)
-    }
-
-  private def partyFilter[Context](allowedParties: Set[String]) =
-    Flow[Ctx[Context, CommandSubmission]].map { elem =>
-      val commands = elem.value.commands
-      val effectiveActAs = CommandsValidator.effectiveSubmitters(commands).actAs
-      if (effectiveActAs.subsetOf(allowedParties)) elem
-      else
-        throw new IllegalArgumentException(
-          s"Attempted submission and tracking of command ${commands.commandId} by parties $effectiveActAs while some of those parties are not part of the subscription set $allowedParties."
-        )
-    }
+  ): Future[TrackCommandFlow[Context]] = it.trackCommandsUnbounded(parties, token, ledgerId)
 
   def completionSource(
       parties: Seq[String],
       offset: LedgerOffset,
       token: Option[String] = None,
-      ledgerIdToUse: LedgerId,
-  ): Source[CompletionStreamElement, NotUsed] = {
-    logger.debug(
-      "Connecting to completion service with parties '{}' from offset: '{}'",
-      parties,
-      offset: Any,
-    )
-    CommandCompletionSource(
-      CompletionStreamRequest(
-        ledgerIdToUse.unwrap,
-        applicationId,
-        parties,
-        Some(offset),
-      ),
-      LedgerClient.stub(commandCompletionService, token).completionStream,
-    )
-  }
-
-  private def commandUpdaterFlow[Context](ledgerIdToUse: LedgerId) =
-    Flow[Ctx[Context, CommandSubmission]]
-      .map(_.map { case submission @ CommandSubmission(commands) =>
-        if (LedgerId(commands.ledgerId) != ledgerIdToUse)
-          throw new IllegalArgumentException(
-            s"Failing fast on submission request of command ${commands.commandId} with invalid ledger ID ${commands.ledgerId} (client expected $ledgerIdToUse)"
-          )
-        else if (commands.applicationId != applicationId)
-          throw new IllegalArgumentException(
-            s"Failing fast on submission request of command ${commands.commandId} with invalid application ID ${commands.applicationId} (client expected $applicationId)"
-          )
-        val updatedDeduplicationPeriod = commands.deduplicationPeriod match {
-          case DeduplicationPeriod.Empty =>
-            DeduplicationPeriod.DeduplicationTime(
-              Duration
-                .of(
-                  config.defaultDeduplicationTime.getSeconds,
-                  config.defaultDeduplicationTime.getNano,
-                )
-            )
-          case existing => existing
-        }
-        submission.copy(commands = commands.copy(deduplicationPeriod = updatedDeduplicationPeriod))
-      })
+  ): Source[CompletionStreamElement, NotUsed] =
+    it.completionSource(parties, offset, token, ledgerId)
 
   def submissionFlow[Context](
-      token: Option[String] = None,
-      ledgerIdToUse: LedgerId,
-  ): Flow[Ctx[Context, CommandSubmission], Ctx[Context, Try[Empty]], NotUsed] = {
-    Flow[Ctx[Context, CommandSubmission]]
-      .via(commandUpdaterFlow(ledgerIdToUse))
-      .via(CommandSubmissionFlow[Context](submit(token), config.maxParallelSubmissions))
-  }
+      token: Option[String] = None
+  ): Flow[Ctx[Context, CommandSubmission], Ctx[Context, Try[Empty]], NotUsed] =
+    it.submissionFlow(token, ledgerId)
 
   def getCompletionEnd(
-      token: Option[String] = None,
-      ledgerIdToUse: LedgerId,
-  ): Future[CompletionEndResponse] =
-    LedgerClient
-      .stub(commandCompletionService, token)
-      .completionEnd(
-        CompletionEndRequest(ledgerIdToUse.unwrap)
-      )
-}
-
-/** Enables easy access to command services and high level operations on top of them.
-  *
-  * @param commandSubmissionService gRPC service reference.
-  * @param commandCompletionService gRPC service reference.
-  * @param ledgerId                 Will be applied to submitted commands.
-  * @param applicationId            Will be applied to submitted commands.
-  * @param config                   Options for changing behavior.
-  */
-final class CommandClient(
-    commandSubmissionService: CommandSubmissionServiceStub,
-    commandCompletionService: CommandCompletionServiceStub,
-    val ledgerId: LedgerId,
-    applicationId: String,
-    config: CommandClientConfiguration,
-    logger: Logger = LoggerFactory.getLogger(getClass),
-)(implicit esf: ExecutionSequencerFactory)
-    extends CommandClientWithoutLedgerId(
-      commandSubmissionService,
-      commandCompletionService,
-      applicationId,
-      config,
-      logger,
-    ) {
-
-  /** Submits and tracks a single command. High frequency usage is discouraged as it causes a dedicated completion
-    * stream to be established and torn down.
-    */
-  override def trackSingleCommand(
-      submitRequest: SubmitRequest,
-      token: Option[String] = None,
-      ledgerIdToUse: LedgerId = ledgerId,
-  )(implicit
-      mat: Materializer
-  ): Future[Either[CompletionFailure, CompletionSuccess]] =
-    super.trackSingleCommand(submitRequest, token, ledgerId)
-
-  /** Tracks the results (including timeouts) of incoming commands.
-    * Applies a maximum bound for in-flight commands which have been submitted, but not confirmed through command completions.
-    *
-    * The resulting flow will backpressure if downstream backpressures, independently of the number of in-flight commands.
-    *
-    * @param parties Commands that have a submitting party which is not part of this collection will fail the stream.
-    */
-  override def trackCommands[Context](
-      parties: Seq[String],
-      token: Option[String] = None,
-      ledgerIdToUse: LedgerId = ledgerId,
-  )(implicit
-      ec: ExecutionContext
-  ): Future[TrackCommandFlow[Context]] = super.trackCommands(parties, token, ledgerId)
-
-  /** Tracks the results (including timeouts) of incoming commands.
-    *
-    * The resulting flow will backpressure if downstream backpressures, independently of the number of in-flight commands.
-    *
-    * @param parties Commands that have a submitting party which is not part of this collection will fail the stream.
-    */
-  override def trackCommandsUnbounded[Context](
-      parties: Seq[String],
-      token: Option[String] = None,
-      ledgerIdToUse: LedgerId = ledgerId,
-  )(implicit
-      ec: ExecutionContext
-  ): Future[TrackCommandFlow[Context]] = super.trackCommandsUnbounded(parties, token, ledgerId)
-
-  override def completionSource(
-      parties: Seq[String],
-      offset: LedgerOffset,
-      token: Option[String] = None,
-      ledgerIdToUse: LedgerId = ledgerId,
-  ): Source[CompletionStreamElement, NotUsed] =
-    super.completionSource(parties, offset, token, ledgerId)
-
-  override def submissionFlow[Context](
-      token: Option[String] = None,
-      ledgerIdToUse: LedgerId = ledgerId,
-  ): Flow[Ctx[Context, CommandSubmission], Ctx[Context, Try[Empty]], NotUsed] =
-    super.submissionFlow(token, ledgerId)
-
-  override def getCompletionEnd(
-      token: Option[String] = None,
-      ledgerIdToUse: LedgerId = ledgerId,
-  ): Future[CompletionEndResponse] = super.getCompletionEnd(token, ledgerId)
+      token: Option[String] = None
+  ): Future[CompletionEndResponse] = it.getCompletionEnd(token, ledgerId)
 }
